:sparkles: Add support for changing the game speed

+++ Source/FrameUncap/IFrameUncapStrategy.cs
@@ -7,5 +7,5 @@ public interface IFrameUncapStrategy
     public TimeSpan TargetUpdateElapsedTime { get; protected set; }
     public TimeSpan TargetDrawElapsedTime { get; protected set; }
 
-    public void SetTargetFramerate(int updateFramerate, int drawFramerate);
+    public void SetTargetFramerate(double updateFramerate, double drawFramerate);
 }
